requirements.txt
Django==3.2.0
django-cors-headers==3.11.0
django-filter==2.3.0
djangorestframework==3.13.1
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.2.4
django-sequences==2.6
psycopg2==2.9.6